Fix graphql-java relay metadata packaging
