+++ tests/src/com.graphql-java/graphql-java/19.7/src/test/java/com_graphql_java/graphql_java/GraphQlTests.java
@@ -4,13 +4,17 @@
  * You should have received a copy of the CC0 legalcode along with this
  * work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
  */
-package graphql;
+package com_graphql_java.graphql_java;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
 import java.util.function.Consumer;
 
+import com_graphql_java.graphql_java.starwars.Droid;
+import com_graphql_java.graphql_java.starwars.Human;
+import graphql.ExecutionResult;
+import graphql.GraphQL;
 import graphql.introspection.IntrospectionQuery;
 import graphql.relay.Connection;
 import graphql.relay.DefaultConnection;
@@ -26,15 +30,11 @@ import graphql.schema.idl.RuntimeWiring;
 import graphql.schema.idl.SchemaGenerator;
 import graphql.schema.idl.SchemaParser;
 import graphql.schema.idl.TypeDefinitionRegistry;
-import graphql.starwars.Droid;
-import graphql.starwars.Human;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-/**
- * @author Brian Clozel
- */
+/** Exercises graphql-java through its public API in an isolated test namespace. */
 public class GraphQlTests {
 
 
@@ -77,7 +77,7 @@ public class GraphQlTests {
   }
 
   private GraphQLSchema parseSchema(String schemaFileName, Consumer<RuntimeWiring.Builder> consumer) throws IOException {
-    try (InputStream inputStream = GraphQlTests.class.getResourceAsStream(schemaFileName + ".graphqls")) {
+    try (InputStream inputStream = GraphQlTests.class.getResourceAsStream("/graphql/" + schemaFileName + ".graphqls")) {
       TypeDefinitionRegistry registry = new SchemaParser().parse(inputStream);
       RuntimeWiring.Builder runtimeWiringBuilder = RuntimeWiring.newRuntimeWiring();
       consumer.accept(runtimeWiringBuilder);
