+++ tests/src/com.graphql-java/graphql-java/19.7/src/test/java/com_graphql_java/graphql_java/starwars/Character.java
@@ -4,13 +4,21 @@
  * You should have received a copy of the CC0 legalcode along with this
  * work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
  */
-package graphql.starwars;
+package com_graphql_java.graphql_java.starwars;
+
+import java.util.List;
 
 /**
  * @author Brian Clozel
  */
-public enum Episode {
-  NEWHOPE,
-  EMPIRE,
-  JEDI
+public interface Character {
+
+  Long getId();
+
+  String getName();
+
+  Character getFriends();
+
+  List<Episode> getAppearsIn();
+
 }
