+++ tests/src/com.graphql-java/graphql-java/19.7/src/test/java/com_graphql_java/graphql_java/starwars/Episode.java
@@ -4,21 +4,13 @@
  * You should have received a copy of the CC0 legalcode along with this
  * work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
  */
-package graphql.starwars;
-
-import java.util.List;
+package com_graphql_java.graphql_java.starwars;
 
 /**
  * @author Brian Clozel
  */
-public interface Character {
-
-  Long getId();
-
-  String getName();
-
-  Character getFriends();
-
-  List<Episode> getAppearsIn();
-
+public enum Episode {
+  NEWHOPE,
+  EMPIRE,
+  JEDI
 }
