+++ tests/src/com.graphql-java/graphql-java/19.7/src/test/resources/META-INF/native-image/reachability-metadata.json
@@ -1,104 +1,22 @@
 {
   "reflection" : [
     {
-      "condition" : {
-        "typeReached" : "graphql.relay.Connection"
-      },
-      "type" : "graphql.relay.DefaultConnection",
-      "allDeclaredMethods" : true
-    },
-    {
-      "condition" : {
-        "typeReached" : "graphql.relay.Connection"
-      },
-      "type" : "graphql.relay.DefaultConnectionCursor",
-      "allDeclaredMethods" : true
-    },
-    {
-      "condition" : {
-        "typeReached" : "graphql.relay.Connection"
-      },
-      "type" : "graphql.relay.DefaultEdge",
-      "allDeclaredMethods" : true
-    },
-    {
-      "condition" : {
-        "typeReached" : "graphql.relay.Connection"
-      },
-      "type" : "graphql.relay.DefaultPageInfo",
-      "allDeclaredMethods" : true
-    },
-    {
-      "condition" : {
-        "typeReached" : "graphql.GraphQL"
-      },
-      "type" : "graphql.schema.GraphQLArgument",
-      "allPublicMethods" : true
-    },
-    {
-      "condition" : {
-        "typeReached" : "graphql.GraphQL"
-      },
-      "type" : "graphql.schema.GraphQLDirective",
-      "allPublicMethods" : true
-    },
-    {
-      "condition" : {
-        "typeReached" : "graphql.GraphQL"
-      },
-      "type" : "graphql.schema.GraphQLEnumValueDefinition",
-      "allPublicMethods" : true
-    },
-    {
-      "condition" : {
-        "typeReached" : "graphql.GraphQL"
-      },
-      "type" : "graphql.schema.GraphQLFieldDefinition",
-      "allPublicMethods" : true
-    },
-    {
-      "condition" : {
-        "typeReached" : "graphql.GraphQL"
-      },
-      "type" : "graphql.schema.GraphQLInputObjectField",
-      "allPublicMethods" : true
-    },
-    {
-      "condition" : {
-        "typeReached" : "graphql.GraphQL"
-      },
-      "type" : "graphql.schema.GraphQLOutputType",
-      "allPublicMethods" : true
-    },
-    {
-      "type" : "graphql.starwars.Human",
+      "type" : "com_graphql_java.graphql_java.starwars.Human",
       "allPublicMethods" : true
     }
   ],
   "resources" : [
     {
       "condition" : {
-        "typeReached" : "graphql.GraphQlTests"
+        "typeReached" : "com_graphql_java.graphql_java.GraphQlTests"
       },
       "glob" : "graphql/greeting.graphqls"
     },
     {
       "condition" : {
-        "typeReached" : "graphql.GraphQlTests"
+        "typeReached" : "com_graphql_java.graphql_java.GraphQlTests"
       },
       "glob" : "graphql/starwars.graphqls"
-    },
-    {
-      "condition" : {
-        "typeReached" : "org.junit.platform.reporting.legacy.xml.XmlReportWriter"
-      },
-      "glob" : "META-INF/services/javax.xml.stream.XMLOutputFactory"
-    },
-    {
-      "condition" : {
-        "typeReached" : "org.junit.platform.reporting.legacy.xml.XmlReportWriter"
-      },
-      "glob" : "META-INF/services/java.net.spi.InetAddressResolverProvider"
     }
   ]
 }
